fix: 순환 참조 문제 및 강좌 정렬 문제 해결
